Module 2/week-10/flask-coursemodel/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marcelo Coelho <92119579+MCoelho222@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.1"
python-dotenv = "^0.20.0"
Flask-SQLAlchemy = "^2.5.1"
psycopg2 = "^2.9.3"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.28.1"
Flask-Migrate = "^3.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
